// src/lib/api/models/createChecklistBodyOne.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Vistoria API
 * Backend do app de vistoria. Auth por email-OTP + organizations (Better Auth em /api/auth). Projetos pertencem à organização; recursos de vistoria são escopados por :projectId no path.
 * OpenAPI spec version: 1.0.0
 */
import type { CreateChecklistBodyOneItemsItem } from './createChecklistBodyOneItemsItem';
import type { CreateChecklistBodyOneMetadata } from './createChecklistBodyOneMetadata';
import type { CreateChecklistBodyOneOptionsItem } from './createChecklistBodyOneOptionsItem';

export type CreateChecklistBodyOne = {
  id?: string;
  title: string;
  source?: string | null;
  tagsIds?: string[];
  options?: CreateChecklistBodyOneOptionsItem[];
  items?: CreateChecklistBodyOneItemsItem[];
  metadata?: CreateChecklistBodyOneMetadata;
};